+++ CHANGES.md
@@ -4,6 +4,7 @@
 * Created export dialog and moved export settings to this window.
 * Improved appearance of Settings dialog.
 * Updated Tiny File Dialogs to 3.2.0. Resolves #45
+* Added Style property type to Custom Classes. Resolves #46
 
 ### Skin Composer Version 25 ###
 * Fixed exception when opening an old project.